update unit test for rocket_delete_licence_data_file
